+++ CHANGELOG.md
@@ -5,11 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.12.0] - Unreleased
+## Unreleased
 
 ### Added
 
 - Added `Widget.disabled` https://github.com/Textualize/textual/pull/1785
+
+### Fixed
+
+- DataTable scrollbars resize correctly when header is toggled https://github.com/Textualize/textual/pull/1803
 
 ## [0.11.0] - 2023-02-15